Show AddUserSecrets without CreateDefaultBuilder
