--- aspnetcore/security/app-secrets/samples/1.x/UserSecrets/Models/MovieSettings.cs
@@ -1,11 +1,0 @@
-﻿namespace UserSecrets.Models
-{
-    #region snippet_MovieSettingsClass
-    public class MovieSettings
-    {
-        public string ConnectionString { get; set; }
-
-        public string ServiceApiKey { get; set; }
-    }
-    #endregion
-}
