--- aspnetcore/security/app-secrets/samples/1.x/UserSecrets/Program.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.AspNetCore.Hosting;
-using System.IO;
-
-namespace UserSecrets
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .UseKestrel()
-                .UseContentRoot(Directory.GetCurrentDirectory())
-                .UseIISIntegration()
-                .UseStartup<Startup>()
-                .Build();
-
-            host.Run();
-        }
-    }
-}
