--- aspnetcore/security/app-secrets/samples/1.x/UserSecrets/Startup2.cs
@@ -1,53 +1,0 @@
-﻿#if never
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using System.Data.SqlClient;
-
-namespace UserSecrets
-{
-    #region snippet_StartupClass
-    public class Startup
-    {
-        private string _connection = null;
-        
-        public Startup(IHostingEnvironment env)
-        {
-            var builder = new ConfigurationBuilder()
-                .SetBasePath(env.ContentRootPath)
-                .AddJsonFile("appsettings.json",
-                             optional: false,
-                             reloadOnChange: true)
-                .AddEnvironmentVariables();
-
-            if (env.IsDevelopment())
-            {
-                builder.AddUserSecrets<Startup>();
-            }
-
-            Configuration = builder.Build();
-        }
-
-        public IConfigurationRoot Configuration { get; }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            var builder = new SqlConnectionStringBuilder(
-                Configuration.GetConnectionString("Movies"));
-            builder.Password = Configuration["DbPassword"];
-            _connection = builder.ConnectionString;
-        }
-
-        public void Configure(IApplicationBuilder app)
-        {
-            app.Run(async (context) =>
-            {
-                await context.Response.WriteAsync($"DB Connection: {_connection}");
-            });
-        }
-    }
-    #endregion snippet_StartupClass
-}
-#endif
