--- aspnetcore/security/app-secrets/samples/1.x/UserSecrets/UserSecrets.csproj
@@ -1,22 +1,0 @@
-﻿<!-- <snippet_CsprojFile> -->
-<Project Sdk="Microsoft.NET.Sdk.Web">
-  <!-- <snippet_PropertyGroup> -->
-  <PropertyGroup>
-    <TargetFramework>netcoreapp1.1</TargetFramework>
-    <UserSecretsId>1242d6d6-9df3-4031-b031-d9b27d13c25a</UserSecretsId>
-  </PropertyGroup>
-  <!-- </snippet_PropertyGroup> -->
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore" 
-                      Version="1.1.6" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" 
-                      Version="1.1.2" />
-    <PackageReference Include="System.Data.SqlClient" 
-                      Version="4.5.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.Extensions.SecretManager.Tools" 
-                            Version="1.0.1" />
-  </ItemGroup>
-</Project>
-<!-- </snippet_CsprojFile> -->
