--- aspnetcore/security/app-secrets/samples/1.x/UserSecrets/appsettings.json
@@ -1,5 +1,0 @@
-﻿{
-  "ConnectionStrings": {
-    "Movies": "Server=(localdb)\\mssqllocaldb;Database=Movie-1;User Id=johndoe;MultipleActiveResultSets=true"
-  }
-}
