+++ aspnetcore/security/app-secrets/samples/2.x/UserSecrets/Startup3.cs
@@ -1,5 +1,6 @@
 ﻿#if never
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -11,10 +12,24 @@ namespace UserSecrets
     {
         private string _moviesApiKey = null;
 
-        public Startup(IConfiguration configuration)
+        #region snippet_StartupConstructor
+        public Startup(IHostingEnvironment env)
         {
-            Configuration = configuration;
+            var builder = new ConfigurationBuilder()
+                .SetBasePath(env.ContentRootPath)
+                .AddJsonFile("appsettings.json", 
+                             optional: false, 
+                             reloadOnChange: true)
+                .AddEnvironmentVariables();
+
+            if (env.IsDevelopment())
+            {
+                builder.AddUserSecrets<Startup>();
+            }
+
+            Configuration = builder.Build();
         }
+        #endregion
 
         public IConfiguration Configuration { get; }
