+++ aspnetcore/security/app-secrets/samples/3.x/UserSecrets/Program2.cs
@@ -1,0 +1,29 @@
+﻿#if never
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
+
+namespace UserSecrets
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            #region snippet_Host
+            var host = new HostBuilder()
+                .ConfigureAppConfiguration((hostContext, builder) =>
+                {
+                    // Add other providers for JSON, etc.
+
+                    if (hostContext.HostingEnvironment.IsDevelopment())
+                    {
+                        builder.AddUserSecrets<Program>();
+                    }
+                })
+                .Build();
+                #endregion
+            
+            host.Run();
+        }
+    }
+}
+#endif
